refactor: move enrollment classes out to their own modules
