+++ benefits/enrollment/switchio.py
@@ -1,0 +1,8 @@
+class SwitchioEnrollment:
+    @staticmethod
+    def initiate_tokenization(request):
+        pass
+
+    @staticmethod
+    def enroll(request):
+        pass
