+++ tests/pytest/enrollment/test_enrollment.py
@@ -6,7 +6,8 @@ from littlepay.api.funding_sources import FundingSourceResponse
 from littlepay.api.groups import GroupFundingSourceResponse
 from requests import HTTPError
 
-from benefits.enrollment.enrollment import Status, LittlepayEnrollment
+from benefits.enrollment.enrollment import Status
+from benefits.enrollment.littlepay import LittlepayEnrollment
 
 
 @pytest.fixture
@@ -91,7 +92,7 @@ def test_calculate_expiry():
 def test_calculate_expiry_specific_date(mocker):
     expiration_days = 14
     mocker.patch(
-        "benefits.enrollment.enrollment.timezone.now",
+        "benefits.enrollment.littlepay.timezone.now",
         return_value=timezone.make_aware(
             value=timezone.datetime(2024, 3, 1, 13, 37, 11, 5), timezone=timezone.get_fixed_timezone(offset=0)
         ),
@@ -109,7 +110,7 @@ def test_is_expired_expiry_date_is_in_the_past(mocker):
 
     # mock datetime of "now" to be specific date for testing
     mocker.patch(
-        "benefits.enrollment.enrollment.timezone.now",
+        "benefits.enrollment.littlepay.timezone.now",
         return_value=timezone.make_aware(timezone.datetime(2024, 1, 1, 10, 30), timezone.get_default_timezone()),
     )
 
@@ -121,7 +122,7 @@ def test_is_expired_expiry_date_is_in_the_future(mocker):
 
     # mock datetime of "now" to be specific date for testing
     mocker.patch(
-        "benefits.enrollment.enrollment.timezone.now",
+        "benefits.enrollment.littlepay.timezone.now",
         return_value=timezone.make_aware(timezone.datetime(2024, 1, 1, 11, 5), timezone.get_default_timezone()),
     )
 
@@ -133,7 +134,7 @@ def test_is_expired_expiry_date_equals_now(mocker):
 
     # mock datetime of "now" to be specific date for testing
     mocker.patch(
-        "benefits.enrollment.enrollment.timezone.now",
+        "benefits.enrollment.littlepay.timezone.now",
         return_value=timezone.make_aware(timezone.datetime(2024, 1, 1, 13, 37), timezone.get_default_timezone()),
     )
 
@@ -146,7 +147,7 @@ def test_is_within_enrollment_window_True(mocker):
 
     # mock datetime of "now" to be specific date for testing
     mocker.patch(
-        "benefits.enrollment.enrollment.timezone.now",
+        "benefits.enrollment.littlepay.timezone.now",
         return_value=timezone.make_aware(timezone.datetime(2023, 2, 15, 15, 30), timezone=timezone.get_default_timezone()),
     )
 
@@ -163,7 +164,7 @@ def test_is_within_enrollment_window_before_window(mocker):
 
     # mock datetime of "now" to be specific date for testing
     mocker.patch(
-        "benefits.enrollment.enrollment.timezone.now",
+        "benefits.enrollment.littlepay.timezone.now",
         return_value=timezone.make_aware(timezone.datetime(2023, 1, 15, 15, 30), timezone=timezone.get_default_timezone()),
     )
 
@@ -180,7 +181,7 @@ def test_is_within_enrollment_window_after_window(mocker):
 
     # mock datetime of "now" to be specific date for testing
     mocker.patch(
-        "benefits.enrollment.enrollment.timezone.now",
+        "benefits.enrollment.littlepay.timezone.now",
         return_value=timezone.make_aware(timezone.datetime(2023, 3, 15, 15, 30), timezone=timezone.get_default_timezone()),
     )
 
@@ -197,7 +198,7 @@ def test_is_within_enrollment_window_equal_reenrollment_date(mocker):
 
     # mock datetime of "now" to be specific date for testing
     mocker.patch(
-        "benefits.enrollment.enrollment.timezone.now",
+        "benefits.enrollment.littlepay.timezone.now",
         return_value=enrollment_reenrollment_date,
     )
 
@@ -214,7 +215,7 @@ def test_is_within_enrollment_window_equal_expiry_date(mocker):
 
     # mock datetime of "now" to be specific date for testing
     mocker.patch(
-        "benefits.enrollment.enrollment.timezone.now",
+        "benefits.enrollment.littlepay.timezone.now",
         return_value=expiry_date,
     )
 
@@ -234,7 +235,7 @@ def test_enroll_system_error(
     app_request,
     card_token,
 ):
-    mock_client_cls = mocker.patch("benefits.enrollment.enrollment.Client")
+    mock_client_cls = mocker.patch("benefits.enrollment.littlepay.Client")
     mock_client = mock_client_cls.return_value
 
     mock_error = {"message": "Mock error message"}
@@ -256,7 +257,7 @@ def test_enroll_exception_http_error_400(
     app_request,
     card_token,
 ):
-    mock_client_cls = mocker.patch("benefits.enrollment.enrollment.Client")
+    mock_client_cls = mocker.patch("benefits.enrollment.littlepay.Client")
     mock_client = mock_client_cls.return_value
 
     mock_error = {"message": "Mock error message"}
@@ -281,7 +282,7 @@ def test_enroll_exception_non_http_error(
     app_request,
     card_token,
 ):
-    mock_client_cls = mocker.patch("benefits.enrollment.enrollment.Client")
+    mock_client_cls = mocker.patch("benefits.enrollment.littlepay.Client")
     mock_client = mock_client_cls.return_value
 
     mock_client.link_concession_group_funding_source.side_effect = Exception("some other exception")
@@ -302,12 +303,12 @@ def test_enroll_success_flow_does_not_support_expiration_customer_already_enroll
     mocked_funding_source,
     mocked_group_funding_source_no_expiry,
 ):
-    mock_client_cls = mocker.patch("benefits.enrollment.enrollment.Client")
+    mock_client_cls = mocker.patch("benefits.enrollment.littlepay.Client")
     mock_client = mock_client_cls.return_value
     mock_client.get_funding_source_by_token.return_value = mocked_funding_source
 
     mocker.patch(
-        "benefits.enrollment.enrollment.LittlepayEnrollment._get_group_funding_source",
+        "benefits.enrollment.littlepay.LittlepayEnrollment._get_group_funding_source",
         return_value=mocked_group_funding_source_no_expiry,
     )
 
@@ -328,7 +329,7 @@ def test_enroll_success_flow_does_not_support_expiration_no_expiry(
     card_token,
     mocked_funding_source,
 ):
-    mock_client_cls = mocker.patch("benefits.enrollment.enrollment.Client")
+    mock_client_cls = mocker.patch("benefits.enrollment.littlepay.Client")
     mock_client = mock_client_cls.return_value
     mock_client.get_funding_source_by_token.return_value = mocked_funding_source
 
@@ -351,7 +352,7 @@ def test_enroll_success_flow_supports_expiration(
     mocked_funding_source,
     mocked_session_enrollment_expiry,
 ):
-    mock_client_cls = mocker.patch("benefits.enrollment.enrollment.Client")
+    mock_client_cls = mocker.patch("benefits.enrollment.littlepay.Client")
     mock_client = mock_client_cls.return_value
     mock_client.get_funding_source_by_token.return_value = mocked_funding_source
 
@@ -377,12 +378,12 @@ def test_enroll_success_flow_supports_expiration_no_expiry(
     mocked_group_funding_source_no_expiry,
     mocked_session_enrollment_expiry,
 ):
-    mock_client_cls = mocker.patch("benefits.enrollment.enrollment.Client")
+    mock_client_cls = mocker.patch("benefits.enrollment.littlepay.Client")
     mock_client = mock_client_cls.return_value
     mock_client.get_funding_source_by_token.return_value = mocked_funding_source
 
     mocker.patch(
-        "benefits.enrollment.enrollment.LittlepayEnrollment._get_group_funding_source",
+        "benefits.enrollment.littlepay.LittlepayEnrollment._get_group_funding_source",
         return_value=mocked_group_funding_source_no_expiry,
     )
 
@@ -407,17 +408,17 @@ def test_enroll_success_flow_supports_expiration_is_expired(
     mocked_group_funding_source_with_expiry,
     mocked_session_enrollment_expiry,
 ):
-    mock_client_cls = mocker.patch("benefits.enrollment.enrollment.Client")
+    mock_client_cls = mocker.patch("benefits.enrollment.littlepay.Client")
     mock_client = mock_client_cls.return_value
     mock_client.get_funding_source_by_token.return_value = mocked_funding_source
 
     # mock that a funding source already exists, doesn't matter what expiry_date is
     mocker.patch(
-        "benefits.enrollment.enrollment.LittlepayEnrollment._get_group_funding_source",
+        "benefits.enrollment.littlepay.LittlepayEnrollment._get_group_funding_source",
         return_value=mocked_group_funding_source_with_expiry,
     )
 
-    mocker.patch("benefits.enrollment.enrollment.LittlepayEnrollment._is_expired", return_value=True)
+    mocker.patch("benefits.enrollment.littlepay.LittlepayEnrollment._is_expired", return_value=True)
 
     status, exception = LittlepayEnrollment.enroll(app_request, card_token)
 
@@ -441,17 +442,17 @@ def test_enroll_success_flow_supports_expiration_is_within_reenrollment_window(
     mocked_group_funding_source_with_expiry,
     mocked_session_enrollment_expiry,
 ):
-    mock_client_cls = mocker.patch("benefits.enrollment.enrollment.Client")
+    mock_client_cls = mocker.patch("benefits.enrollment.littlepay.Client")
     mock_client = mock_client_cls.return_value
     mock_client.get_funding_source_by_token.return_value = mocked_funding_source
 
     # mock that a funding source already exists, doesn't matter what expiry_date is
     mocker.patch(
-        "benefits.enrollment.enrollment.LittlepayEnrollment._get_group_funding_source",
+        "benefits.enrollment.littlepay.LittlepayEnrollment._get_group_funding_source",
         return_value=mocked_group_funding_source_with_expiry,
     )
 
-    mocker.patch("benefits.enrollment.enrollment.LittlepayEnrollment._is_within_reenrollment_window", return_value=True)
+    mocker.patch("benefits.enrollment.littlepay.LittlepayEnrollment._is_within_reenrollment_window", return_value=True)
 
     status, exception = LittlepayEnrollment.enroll(app_request, card_token)
 
@@ -473,18 +474,18 @@ def test_enroll_reenrollment_error(
     mocked_funding_source,
     mocked_group_funding_source_with_expiry,
 ):
-    mock_client_cls = mocker.patch("benefits.enrollment.enrollment.Client")
+    mock_client_cls = mocker.patch("benefits.enrollment.littlepay.Client")
     mock_client = mock_client_cls.return_value
     mock_client.get_funding_source_by_token.return_value = mocked_funding_source
 
     # mock that a funding source already exists, doesn't matter what expiry_date is
     mocker.patch(
-        "benefits.enrollment.enrollment.LittlepayEnrollment._get_group_funding_source",
+        "benefits.enrollment.littlepay.LittlepayEnrollment._get_group_funding_source",
         return_value=mocked_group_funding_source_with_expiry,
     )
 
-    mocker.patch("benefits.enrollment.enrollment.LittlepayEnrollment._is_expired", return_value=False)
-    mocker.patch("benefits.enrollment.enrollment.LittlepayEnrollment._is_within_reenrollment_window", return_value=False)
+    mocker.patch("benefits.enrollment.littlepay.LittlepayEnrollment._is_expired", return_value=False)
+    mocker.patch("benefits.enrollment.littlepay.LittlepayEnrollment._is_within_reenrollment_window", return_value=False)
 
     status, exception = LittlepayEnrollment.enroll(app_request, card_token)
 
@@ -502,13 +503,13 @@ def test_enroll_does_not_support_expiration_has_expiration_date(
     mocked_funding_source,
     mocked_group_funding_source_with_expiry,
 ):
-    mock_client_cls = mocker.patch("benefits.enrollment.enrollment.Client")
+    mock_client_cls = mocker.patch("benefits.enrollment.littlepay.Client")
     mock_client = mock_client_cls.return_value
     mock_client.get_funding_source_by_token.return_value = mocked_funding_source
 
     # mock that a funding source already exists, doesn't matter what expiry_date is
     mocker.patch(
-        "benefits.enrollment.enrollment.LittlepayEnrollment._get_group_funding_source",
+        "benefits.enrollment.littlepay.LittlepayEnrollment._get_group_funding_source",
         return_value=mocked_group_funding_source_with_expiry,
     )
 
@@ -532,7 +533,7 @@ def test_request_card_tokenization_access(mocker, app_request):
     mock_response["access_token"] = "123"
     mock_response["expires_at"] = "2024-01-01T00:00:00"
 
-    mock_client_cls = mocker.patch("benefits.enrollment.enrollment.Client")
+    mock_client_cls = mocker.patch("benefits.enrollment.littlepay.Client")
     mock_client = mock_client_cls.return_value
     mock_client.request_card_tokenization_access.return_value = mock_response
 
@@ -549,7 +550,7 @@ def test_request_card_tokenization_access(mocker, app_request):
 @pytest.mark.django_db
 @pytest.mark.usefixtures("mocked_session_agency")
 def test_request_card_tokenization_access_system_error(mocker, app_request):
-    mock_client_cls = mocker.patch("benefits.enrollment.enrollment.Client")
+    mock_client_cls = mocker.patch("benefits.enrollment.littlepay.Client")
     mock_client = mock_client_cls.return_value
 
     mock_error = {"message": "Mock error message"}
@@ -571,7 +572,7 @@ def test_request_card_tokenization_access_system_error(mocker, app_request):
 @pytest.mark.django_db
 @pytest.mark.usefixtures("mocked_session_agency")
 def test_request_card_tokenization_access_http_400_error(mocker, app_request):
-    mock_client_cls = mocker.patch("benefits.enrollment.enrollment.Client")
+    mock_client_cls = mocker.patch("benefits.enrollment.littlepay.Client")
     mock_client = mock_client_cls.return_value
 
     mock_error = {"message": "Mock error message"}
@@ -593,7 +594,7 @@ def test_request_card_tokenization_access_http_400_error(mocker, app_request):
 @pytest.mark.django_db
 @pytest.mark.usefixtures("mocked_session_agency")
 def test_request_card_tokenization_access_exception(mocker, app_request):
-    mock_client_cls = mocker.patch("benefits.enrollment.enrollment.Client")
+    mock_client_cls = mocker.patch("benefits.enrollment.littlepay.Client")
     mock_client = mock_client_cls.return_value
 
     exception = Exception("some exception")
